fix: adjust components
