+++ src/pages/prospect/components/TableExtraordinaryInstallment/Actions/styles.ts
