+++ src/pages/prospect/components/TableExtraordinaryInstallment/Detail/index.tsx
