+++ src/pages/prospect/components/TableExtraordinaryInstallment/config/index.tsx
@@ -1,0 +1,45 @@
+import { currencyFormat } from "@src/utils/formatData/currency";
+import { Detail } from "../Detail";
+import { IHeaders } from "@components/modals/ExtraordinaryPaymentModal/types";
+
+export const rowsVisbleMobile = ["datePayment", "value", "actions"];
+
+export const rowsActions = [
+  {
+    label: "Acciones",
+    key: "actions",
+    container: (
+      onClickDetails: () => void,
+      onClickEdit: () => void,
+      onClickEliminate: () => void
+    ) => {
+      return (
+        <Detail
+          onClickDetails={onClickDetails}
+          onClickEdit={onClickEdit}
+          onClickEliminate={onClickEliminate}
+        />
+      );
+    },
+  },
+];
+
+export const headersTableExtraordinaryInstallment: IHeaders[] = [
+  {
+    label: "Fecha de pago",
+    key: "datePayment",
+  },
+  {
+    label: "Valor",
+    key: "value",
+    mask: (value: string | number) => {
+      return currencyFormat(value as number);
+    },
+  },
+  {
+    label: "Medio de pago",
+    key: "paymentMethod",
+  },
+];
+
+export const pageLength = 5;
