+++ src/pages/prospect/components/TableExtraordinaryInstallment/index.tsx
