+++ src/pages/prospect/components/TableExtraordinaryInstallment/stories/TableExtraordinaryPayment.stories.tsx
