+++ src/pages/prospect/components/TableExtraordinaryInstallment/stories/props.ts
